fix: bot deleting the unlock message for lockdown command

+++ bot/cogs/moderation/moderation.py
@@ -42,7 +42,7 @@ class Moderation(commands.Cog):
     @commands.Cog.listener()
     async def on_message(self, message: discord.Message) -> None:
         locked_channels = await self.bot.database.channel_lock.get_locked_channels(message.guild.id)
-        if message.channel.id in locked_channels:
+        if message.channel.id in locked_channels and message.author.id != self.bot.application_id:
             await message.delete()
 
     @moderation.command(name="purge", description="Purge x messages from a channel")
